refactor(lpa-questionnaire-web-app): rename and relocate files for add supplementary documents
